refactor: extract MediaCard, replace inline toast, clean dead waveform code
- Extract inline MediaCard from MediaTab into @/components/ui/MediaCard
- Replace inline toast markup in MediaTab with <SuccessToast> component
- Remove dead canvas/WebAudio code from AudioWaveform (no <canvas> rendered)
- Anchor MediaCardWaveform bars to bottom of canvas
